Make single indexes file and shorten table names
