--- src/chain/redb/indexes/address_stake_part.rs
@@ -1,31 +1,0 @@
-use ::redb::{MultimapTableDefinition, ReadTransaction};
-use std::hash::{DefaultHasher, Hash as _, Hasher};
-
-use crate::model::BlockSlot;
-
-type Error = crate::chain::ChainError;
-
-pub struct AddressStakePartApproxIndexTable;
-impl AddressStakePartApproxIndexTable {
-    pub const DEF: MultimapTableDefinition<'static, u64, u64> =
-        MultimapTableDefinition::new("addressstakepartapproxindextable");
-
-    pub fn compute_key(address_stake_part: &Vec<u8>) -> u64 {
-        let mut hasher = DefaultHasher::new();
-        address_stake_part.hash(&mut hasher);
-        hasher.finish()
-    }
-
-    pub fn get_by_address_stake_part(
-        rx: &ReadTransaction,
-        address_stake_part: &[u8],
-    ) -> Result<Vec<BlockSlot>, Error> {
-        let table = rx.open_multimap_table(Self::DEF)?;
-        let key = Self::compute_key(&address_stake_part.to_vec());
-        let mut out = vec![];
-        for slot in table.get(key)? {
-            out.push(slot?.value());
-        }
-        Ok(out)
-    }
-}
