--- src/chain/redb/indexes/tx_hash.rs
@@ -1,28 +1,0 @@
-use ::redb::{MultimapTableDefinition, ReadTransaction};
-use std::hash::{DefaultHasher, Hash as _, Hasher};
-
-use crate::model::BlockSlot;
-
-type Error = crate::chain::ChainError;
-
-pub struct TxHashApproxIndexTable;
-impl TxHashApproxIndexTable {
-    pub const DEF: MultimapTableDefinition<'static, u64, u64> =
-        MultimapTableDefinition::new("txsapproxindex");
-
-    pub fn compute_key(tx_hash: &Vec<u8>) -> u64 {
-        let mut hasher = DefaultHasher::new();
-        tx_hash.hash(&mut hasher);
-        hasher.finish()
-    }
-
-    pub fn get_by_tx_hash(rx: &ReadTransaction, tx_hash: &[u8]) -> Result<Vec<BlockSlot>, Error> {
-        let table = rx.open_multimap_table(Self::DEF)?;
-        let key = Self::compute_key(&tx_hash.to_vec());
-        let mut out = vec![];
-        for slot in table.get(key)? {
-            out.push(slot?.value());
-        }
-        Ok(out)
-    }
-}
